+++ packages/apollo-utilities/src/directives.ts
@@ -69,9 +69,7 @@ export function shouldInclude(
       // means it has to be a variable value if this is a valid @skip or @include directive
       if (ifValue.kind !== 'Variable') {
         throw new Error(
-          `Argument for the @${
-            directiveName
-          } directive must be a variable or a boolean value.`,
+          `Argument for the @${directiveName} directive must be a variable or a boolean value.`,
         );
       } else {
         evaledValue = variables[(ifValue as VariableNode).name.value];