+++ packages/apollo-cache-inmemory/src/readFromStore.ts
@@ -8,6 +8,7 @@ import {
   getQueryDefinition,
   isJsonValue,
   isIdValue,
+  toIdValue,
   getStoreKeyName,
 } from 'apollo-utilities';
 import { Cache } from 'apollo-cache';
@@ -71,19 +72,22 @@ const readStoreResolver: Resolver = (
 
   if (typeof fieldValue === 'undefined') {
     if (
-      context.cacheResolvers &&
+      context.cacheRedirects &&
       obj &&
       (obj.__typename || objId === 'ROOT_QUERY')
     ) {
       const typename = obj.__typename || 'Query';
 
       // Look for the type in the custom resolver map
-      const type = context.cacheResolvers[typename];
+      const type = context.cacheRedirects[typename];
       if (type) {
         // Look for the field in the custom resolver map
         const resolver = type[fieldName];
         if (resolver) {
-          fieldValue = resolver(obj, args);
+          fieldValue = resolver(obj, args, {
+            getCacheKey: (obj: { __typename: string; id: string | number }) =>
+              toIdValue(context.dataIdFromObject(obj)),
+          });
         }
       }
     }
@@ -161,7 +165,8 @@ export function diffQueryAgainstStore<T>({
     // Global settings
     store,
     returnPartialData,
-    cacheResolvers: (config && config.cacheResolvers) || {},
+    dataIdFromObject: (config && config.dataIdFromObject) || null,
+    cacheRedirects: (config && config.cacheRedirects) || {},
     // Flag set during execution
     hasMissingField: false,
   };