+++ packages/apollo-utilities/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ### vNext
+- Map coverage to original source
 
 ### 1.1.3
 - dependency updates