+++ packages/apollo-utilities/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### vNext
 - Map coverage to original source
+- Don't `deepFreeze` in development/test environments if ES6 symbols are polyfilled [PR#3082](https://github.com/apollographql/apollo-client/pull/3082)
 
 ### 1.1.3
 - dependency updates