+++ packages/apollo-utilities/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 
 ### vNext
+- Map coverage to original source
+- Added ability to include or ignore fragments in `getDirectivesFromDocument` [PR#3010](https://github.com/apollographql/apollo-client/pull/3010)
+
+### 1.1.3
+- dependency updates
+- Added getDirectivesFromDocument utility function
+[PR#2974](https://github.com/apollographql/apollo-client/pull/2974)
+
+### 1.1.2
+- Add client, rest, and export directives to list of known directives [PR#2949](https://github.com/apollographql/apollo-client/pull/2949)
 
 ### 1.1.1
 - Fix typo in error message for invalid argument being passed to @skip or @include directives [PR#2867](https://github.com/apollographql/apollo-client/pull/2867)